chore: delete old table tree component
