+++ packages/frontend/src/components/Explorer/ExploreTree/TableTree/index.tsx
@@ -1,32 +1,52 @@
-import { Classes, Collapse, Colors, Icon, Text } from '@blueprintjs/core';
-import { Tooltip2 } from '@blueprintjs/popover2';
+import {
+    Button,
+    Classes,
+    Collapse,
+    Colors,
+    Icon,
+    Menu,
+    MenuItem,
+    PopoverPosition,
+    Text,
+} from '@blueprintjs/core';
+import { Popover2, Tooltip2 } from '@blueprintjs/popover2';
 import {
     AdditionalMetric,
     CompiledTable,
     Dimension,
+    DimensionType,
+    fieldId,
     friendlyName,
     getItemId,
     hasIntersection,
     hexToRGB,
     isAdditionalMetric,
     isDimension,
+    isFilterableField,
     Metric,
+    MetricType,
+    Source,
     TimeInterval,
 } from '@lightdash/common';
-import React, { FC, useEffect } from 'react';
+import React, { FC, ReactNode, useCallback, useEffect, useMemo } from 'react';
 import { useToggle } from 'react-use';
 import styled from 'styled-components';
 import { getItemBgColor } from '../../../../hooks/useColumns';
-import { TrackSection } from '../../../../providers/TrackingProvider';
-import { SectionName } from '../../../../types/Events';
+import { useFilters } from '../../../../hooks/useFilters';
+import { useExplorer } from '../../../../providers/ExplorerProvider';
+import {
+    TrackSection,
+    useTracking,
+} from '../../../../providers/TrackingProvider';
+import { EventName, SectionName } from '../../../../types/Events';
 import HighlightedText from '../../../common/HighlightedText';
 import DocumentationHelpButton from '../../../DocumentationHelpButton';
 import {
-    CustomMetricButtons,
-    getItemIconName,
-    NodeItemButtons,
-} from './TableTree';
-import { TooltipContent } from './TableTree.styles';
+    ItemOptions,
+    Placeholder,
+    TooltipContent,
+    WarningIcon,
+} from './TableTree.styles';
 import { TableTreeProvider, useTableTreeContext } from './TableTreeProvider';
 
 export type Node = {
@@ -52,6 +72,266 @@ const timeIntervalSort = [
     TimeInterval.MONTH,
     TimeInterval.YEAR,
 ];
+
+export const getItemIconName = (type: DimensionType | MetricType) => {
+    switch (type) {
+        case DimensionType.STRING || MetricType.STRING:
+            return 'citation';
+        case DimensionType.NUMBER || MetricType.NUMBER:
+            return 'numerical';
+        case DimensionType.DATE || MetricType.DATE:
+            return 'calendar';
+        case DimensionType.BOOLEAN || MetricType.BOOLEAN:
+            return 'segmented-control';
+        case DimensionType.TIMESTAMP:
+            return 'time';
+        default:
+            return 'numerical';
+    }
+};
+
+const getCustomMetricType = (type: DimensionType): MetricType[] => {
+    switch (type) {
+        case DimensionType.STRING:
+        case DimensionType.TIMESTAMP:
+        case DimensionType.DATE:
+            return [
+                MetricType.COUNT_DISTINCT,
+                MetricType.COUNT,
+                MetricType.MIN,
+                MetricType.MAX,
+            ];
+
+        case DimensionType.NUMBER:
+            return [
+                MetricType.MIN,
+                MetricType.MAX,
+                MetricType.SUM,
+                MetricType.AVERAGE,
+                MetricType.COUNT_DISTINCT,
+                MetricType.COUNT,
+            ];
+        case DimensionType.BOOLEAN:
+            return [MetricType.COUNT_DISTINCT, MetricType.COUNT];
+        default:
+            return [];
+    }
+};
+
+const NodeItemButtons: FC<{
+    node: Metric | Dimension;
+    onOpenSourceDialog: (source: Source) => void;
+    isHovered: boolean;
+    isSelected: boolean;
+}> = ({ node, onOpenSourceDialog, isHovered, isSelected }) => {
+    const { isFilteredField, addFilter } = useFilters();
+    const isFiltered = isFilteredField(node);
+    const { track } = useTracking();
+
+    const {
+        actions: { addAdditionalMetric },
+    } = useExplorer();
+
+    const createCustomMetric = useCallback(
+        (dimension: Dimension, type: MetricType) => {
+            const shouldCopyFormatting = [
+                MetricType.AVERAGE,
+                MetricType.SUM,
+                MetricType.MIN,
+                MetricType.MAX,
+            ].includes(type);
+            const format =
+                shouldCopyFormatting && dimension.format
+                    ? { format: dimension.format }
+                    : {};
+
+            const defaultRound =
+                type === MetricType.AVERAGE ? { round: 2 } : {};
+            const round =
+                shouldCopyFormatting && dimension.round
+                    ? { round: dimension.round }
+                    : defaultRound;
+
+            addAdditionalMetric({
+                name: `${dimension.name}_${type}`,
+                label: `${friendlyName(type)} of ${dimension.label}`,
+                table: dimension.table,
+                sql: dimension.sql,
+                description: `${friendlyName(type)} of ${
+                    dimension.label
+                } on the table ${dimension.tableLabel}`,
+                type,
+                ...format,
+                ...round,
+            });
+        },
+        [addAdditionalMetric],
+    );
+
+    const menuItems = useMemo<ReactNode[]>(() => {
+        const items: ReactNode[] = [];
+        if (node.source) {
+            items.push(
+                <MenuItem
+                    key="source"
+                    icon={<Icon icon="console" />}
+                    text="Source"
+                    onClick={(e) => {
+                        if (node.source === undefined) {
+                            return;
+                        }
+                        e.stopPropagation();
+                        onOpenSourceDialog(node.source);
+                    }}
+                />,
+            );
+        }
+        if (isFilterableField(node)) {
+            items.push(
+                <MenuItem
+                    key="filter"
+                    icon="filter"
+                    text="Add filter"
+                    onClick={(e) => {
+                        track({
+                            name: EventName.ADD_FILTER_CLICKED,
+                        });
+                        e.stopPropagation();
+                        addFilter(node, undefined);
+                    }}
+                />,
+            );
+        }
+
+        if (
+            node.fieldType === 'dimension' &&
+            getCustomMetricType(node.type).length > 0
+        ) {
+            items.push(
+                <MenuItem
+                    key="custommetric"
+                    icon="clean"
+                    text="Add custom metric"
+                >
+                    {getCustomMetricType(node.type)?.map((metric) => (
+                        <MenuItem
+                            key={metric}
+                            text={friendlyName(metric)}
+                            onClick={(e) => {
+                                e.stopPropagation();
+                                track({
+                                    name: EventName.ADD_CUSTOM_METRIC_CLICKED,
+                                });
+                                createCustomMetric(node, metric);
+                            }}
+                        />
+                    ))}
+                </MenuItem>,
+            );
+        }
+        return items;
+    }, [addFilter, createCustomMetric, node, onOpenSourceDialog, track]);
+
+    return (
+        <ItemOptions>
+            {isFiltered && <Icon icon="filter" />}
+            {node.hidden && (
+                <Tooltip2 content="This field has been hidden in the dbt project. It's recommend to remove it from the query">
+                    <WarningIcon icon={'warning-sign'} intent="warning" />
+                </Tooltip2>
+            )}
+            {menuItems.length > 0 && (isHovered || isSelected) && (
+                <Popover2
+                    content={<Menu>{menuItems}</Menu>}
+                    autoFocus={false}
+                    position={PopoverPosition.BOTTOM_LEFT}
+                    minimal
+                    lazy
+                    interactionKind="click"
+                    renderTarget={({ isOpen, ref, ...targetProps }) => (
+                        <Tooltip2 content="View options">
+                            <Button
+                                {...targetProps}
+                                elementRef={ref === null ? undefined : ref}
+                                icon="more"
+                                minimal
+                                onClick={(e) => {
+                                    (targetProps as any).onClick(e);
+                                    e.stopPropagation();
+                                }}
+                            />
+                        </Tooltip2>
+                    )}
+                />
+            )}
+
+            {isFiltered && !isHovered && !isSelected && <Placeholder />}
+        </ItemOptions>
+    );
+};
+
+const CustomMetricButtons: FC<{
+    node: AdditionalMetric;
+    isHovered: boolean;
+    isSelected: boolean;
+}> = ({ node, isHovered, isSelected }) => {
+    const { track } = useTracking();
+
+    const {
+        actions: { removeAdditionalMetric },
+    } = useExplorer();
+
+    const menuItems = useMemo<ReactNode[]>(() => {
+        return [
+            <MenuItem
+                key="custommetric"
+                icon="delete"
+                text="Remove custom metric"
+                onClick={(e) => {
+                    e.stopPropagation();
+                    track({
+                        name: EventName.REMOVE_CUSTOM_METRIC_CLICKED,
+                    });
+                    removeAdditionalMetric(fieldId(node));
+                }}
+            />,
+        ];
+    }, [removeAdditionalMetric, node, track]);
+
+    return (
+        <div
+            style={{
+                display: 'inline-flex',
+                gap: '10px',
+            }}
+        >
+            {menuItems.length > 0 && (isHovered || isSelected) && (
+                <Popover2
+                    content={<Menu>{menuItems}</Menu>}
+                    autoFocus={false}
+                    position={PopoverPosition.BOTTOM_LEFT}
+                    minimal
+                    lazy
+                    interactionKind="click"
+                    renderTarget={({ isOpen, ref, ...targetProps }) => (
+                        <Tooltip2 content="View options">
+                            <Button
+                                {...targetProps}
+                                elementRef={ref === null ? undefined : ref}
+                                icon="more"
+                                minimal
+                                onClick={(e) => {
+                                    (targetProps as any).onClick(e);
+                                    e.stopPropagation();
+                                }}
+                            />
+                        </Tooltip2>
+                    )}
+                />
+            )}
+        </div>
+    );
+};
 
 const sortNodes =
     (itemsMap: Record<string, Dimension | Metric | AdditionalMetric>) =>
